+++ apps/client/src/app/services/data.service.ts
@@ -146,6 +146,10 @@ export class DataService {
     return this.http.delete<any>(`/api/v1/account/${aId}`);
   }
 
+  public deleteAllOrders() {
+    return this.http.delete<any>(`/api/v1/order/`);
+  }
+
   public deleteOrder(aId: string) {
     return this.http.delete<any>(`/api/v1/order/${aId}`);
   }