+++ libs/ui/src/lib/activities-table/activities-table.component.ts
@@ -49,6 +49,7 @@ export class ActivitiesTableComponent implements OnChanges, OnDestroy, OnInit {
   @Output() activityDeleted = new EventEmitter<string>();
   @Output() activityToClone = new EventEmitter<OrderWithAccount>();
   @Output() activityToUpdate = new EventEmitter<OrderWithAccount>();
+  @Output() deleteAllActivities = new EventEmitter<void>();
   @Output() export = new EventEmitter<string[]>();
   @Output() exportDrafts = new EventEmitter<string[]>();
   @Output() import = new EventEmitter<void>();
@@ -229,6 +230,10 @@ export class ActivitiesTableComponent implements OnChanges, OnDestroy, OnInit {
           return activity.id;
         })
     );
+  }
+
+  public onDeleteAllActivities() {
+    this.deleteAllActivities.emit();
   }
 
   public onImport() {